feat(BluetoothSettings): make modal only show when a bluetooth device wants to pair
The frontend now requests the backend each second while pairing is on.
If the backend says; yes you should display the modal, it will display
the modal but only as long as the pairing countdown.
If the passkey is entered and the countdown is not yet at zero, it
doesnt request the backend anymore because it already has the passkey.